♻️ Move renderer store ipc to central folder
