+++ src/renderer/store-ipc/memory-store.ts
