+++ src/renderer/windows/settings/preload.ts
@@ -2,10 +2,10 @@
 // https://www.electronjs.org/docs/latest/tutorial/process-model#preload-scripts
 
 import { contextBridge, ipcRenderer } from "electron";
-import Store from "../../store/renderer";
+import Store from "../../store-ipc/store";
 import { MemoryStoreSchema, StoreSchema } from "~shared/store/schema";
 import { WindowsEventArguments } from "~shared/types";
-import MemoryStore from "../../memory-store/renderer";
+import MemoryStore from "../../store-ipc/memory-store";
 
 const store = new Store<StoreSchema>();
 const memoryStore = new MemoryStore<MemoryStoreSchema>();
